status, executer: Alligning params to int64
Parsing parameters as int64 will make it easier to compare them in
the checkup stage, should they be needed to analyze results.

+++ pkg/internal/status/status.go
@@ -23,8 +23,8 @@ import kstatus "github.com/kiagnose/kiagnose/kiagnose/status"
 
 type Results struct {
 	TrafficGeneratorMaxDropRate     float64
-	TrafficGeneratorOutErrorPackets int
-	TrafficGeneratorInErrorPackets  int
+	TrafficGeneratorOutErrorPackets int64
+	TrafficGeneratorInErrorPackets  int64
 	DPDKPacketsRxDropped            int64
 	DPDKPacketsTxDropped            int64
 	TrafficGeneratorNode            string
